(FEAT): rename folder and usings from models to entities
