+++ Speedruns.Backend.Tests/Database/DbContextMock.cs
@@ -2,7 +2,7 @@
 
 using Microsoft.EntityFrameworkCore;
 using NSubstitute;
-using Speedruns.Backend.Models;
+using Speedruns.Backend.Entities;
 
 namespace Speedruns.Backend.Tests.Database
 {
