+++ Speedruns.Backend/Controllers/CommentController.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Speedruns.Backend.Models;
+using Speedruns.Backend.Entities;
 using Speedruns.Backend.Interfaces;
 
 namespace Speedruns.Backend.Controllers
