+++ Speedruns.Backend/Controllers/ConsoleController.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 using Microsoft.AspNetCore.Mvc;
-using Speedruns.Backend.Models;
+using Speedruns.Backend.Entities;
 using Speedruns.Backend.Interfaces;
 
 namespace Speedruns.Backend.Controllers
