+++ Speedruns.Backend/Entities/BaseEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace Speedruns.Backend.Models
+﻿namespace Speedruns.Backend.Entities
 {
     public class BaseEntity
     {
