+++ Speedruns.Backend/Entities/SeriesEntity.cs
@@ -1,4 +1,4 @@
-﻿namespace Speedruns.Backend.Models
+﻿namespace Speedruns.Backend.Entities
 {
     public class SeriesEntity : BaseEntity
     {
