+++ Speedruns.Backend/Entities/SpeedrunsContext.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.EntityFrameworkCore;
 
-namespace Speedruns.Backend.Models
+namespace Speedruns.Backend.Entities
 {
     public class SpeedrunsContext: DbContext
     {
